refactor: modified fixture structure and imports
